Upgraded CloudSim Plug from 3.0.1 to 5.1.3
Main differences:
- datacenter.getVmList() is not available anymore and was replaced by datacenter.getVmList.getHostList() -> host.getVmList()
- SimLog.getTasksExecutionInfos(task) is now called via an onFinishListener

The Simulation results before and after the upgrade were compared and are almost identical. The minor differences can be explained by potential changes/bugfixes in CloudSim Plus.

+++ PureEdgeSim/com/mechalikh/pureedgesim/orchestration/Orchestrator.java
@@ -107,7 +107,7 @@ public abstract class Orchestrator {
 					+ " assigned to " + ((EdgeDataCenter) vm.getHost().getDatacenter()).getType() + " vm: " + vm.getId());
 
 			// update history
-			orchestrationHistory.get(vm.getId()).add(task.getId());
+			orchestrationHistory.get((int) vm.getId()).add((int) task.getId());
 		}
 	}
